requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
configparser==5.2.0
discord.py @ git+https://github.com/Rapptz/discord.py@190eeaf19b783f577d035f2067ab3a6900fe07fb
distlib==0.3.4
filelock==3.7.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
orjson==3.7.2
pbr==5.9.0
platformdirs==2.5.2
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.20.0
requests==2.28.0
six==1.16.0
stevedore==3.5.0
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
yarl==1.7.2
